refactor: move constants to own folder
