+++ src/draw/drawLineOnIjs.ts
@@ -4,7 +4,7 @@ import { IJS } from '../IJS';
 import checkProcessable from '../utils/checkProcessable';
 import { getDefaultColor } from '../utils/getDefaultColor';
 import { getOutputImage } from '../utils/getOutputImage';
-import { Point } from '../utils/types';
+import { Point } from '../utils/geometry/points';
 
 export interface DrawLineOnIjsOptions {
   /**
