+++ src/draw/drawLineOnMask.ts
@@ -2,7 +2,7 @@ import { line } from 'bresenham-zingl';
 
 import { Mask } from '../Mask';
 import { maskToOutputMask } from '../utils/getOutputImage';
-import { Point } from '../utils/types';
+import { Point } from '../utils/geometry/points';
 
 export interface DrawLineOnMaskOptions {
   /**
