+++ src/draw/utils/deleteDuplicates.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../utils/types';
+import { Point } from '../../utils/geometry/points';
 
 /**
  * Remove duplicate points from array.
