+++ src/load/__tests__/read.test.ts
@@ -1,6 +1,6 @@
 import { read } from '..';
 import { IJS } from '../../IJS';
-import { ImageColorModel } from '../../utils/colorModels';
+import { ImageColorModel } from '../../utils/constants/colorModels';
 
 test('read existing image', async () => {
   const img = await read(testUtils.getPath('formats/rgba32.png'));
