+++ src/roi/getBorderPoints.ts
@@ -1,4 +1,4 @@
-import { Point } from '../utils/types';
+import { Point } from '../utils/geometry/points';
 
 import { Roi } from './Roi';
